ter_verde_mobile_adapted/requirements.txt
Flask>=2.0
python-dotenv>=1.0
waitress>=2.1
